requirements.txt
appdirs==1.4.3
bottlenose==1.1.2
Django==1.10.6
lxml==3.7.3
mysqlclient==1.3.10
packaging==16.8
pip==9.0.1
pyparsing==2.2.0
python-amazon-simple-product-api==2.2.11
python-dateutil==2.6.0
requests==2.13.0
setuptools==34.3.3
six==1.10.0
wheel==0.29.0
django-bootstrap3>=0.0
